isar-exclude-docs: include "native" package changelogs
So called "native" packages will put the debian changelog into the
package as "changelog" and not "changelog.Debian".
dh_installchangelogs(1)

We have the file from ubuntu and with this change we start doing things
differently. For isar "raw" packages will be "native" so we probably
want that no matter what ubuntu might think. I still started a
discussion on the topic.
https://github.com/tianon/docker-brew-ubuntu-core/issues/216

This was found when working on reproducible builds, for some reason the
changelogs of the raw packages did not differ, because they are not in
the rootfs.

Reported-by: <Venkata.Pyla@toshiba-tsip.com>
Signed-off-by: Henning Schild <henning.schild@siemens.com>